Add files via upload
-How do I interpret user needs and implement them into a program? How does creating “user stories” help with this?
User needs are gathered through a process of requirements determination where a product owner or scrum master will hold an interview with key stakeholders. They will then process these requirements into user stories that will go into a product backlog so that the team can arrange them by priority and user story points. These user story points are a measurement that the team uses to measure how much effort a user story will need to reach a state of done. the team will then use the product backlog to develop the most valuable aspects of the system through sprint iterations.

-How do I approach developing programs? What agile processes do I hope to incorporate into my future development work?
I approach developing programs by analyzing the project and determining what methodology best fits the situation such as either waterfall or agile. If the entire scope of the project is known in advance or it is a relatively small project then the waterfall method may be more beneficial otherwise an agile methodology would be used. If an agile methodology is decided upon then I would follow the agile manifesto and choose a framework such as a scrum framework. Finally, if the project is at scale with many teams needing to collaborate on a single project I must decide how to scale that project such as choosing scrum@scale. 

-What does it mean to be a good team member in software development?
To be a good team member in software development is to be a great communicator through many methods. These methods could be face-to-face, video conferencing, email, documentation, and many other formats. Although face-to-face communication is preferred, it is not always possible which is when video conferencing would be a good alternative, and then lastly by email. Documentation communication is through the creation or manipulation of artifacts by giving clear concise instructions or information about what is being communicated. In the documentation, it is important to think along the lines of creating a document with the intention that once it is turned over anyone could implement the design even if you were going to be unreachable. Another example is when a developer creates or alters code in the product they should again be leaving clear and concise direction except this time explaining why the code exists such as what action it is performing within the system.